Assignment-6/Assignment_6_1/liara/requirements.txt
python-multipart
fastapi
pydantic
psycopg2-binary
uvicorn[standaed]
opencv-python-headless
sqlalchemy
python-dotenv